pyproject.toml
[project]
name = "rss-pocket-feeder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "feedparser>=6.0.11",
    "postgres>=4.0",
    "psycopg>=3.2.3",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
